+++ packages/assets-controllers/src/TokenRatesController.test.ts
@@ -2093,11 +2093,9 @@ describe('TokenRatesController', () => {
               price: 0.002,
             },
           }),
-          validateCurrencySupported: jest.fn().mockReturnValue(
-            false,
-            // Cast used because this method has an assertion in the return
-            // value that I don't know how to type properly with Jest's mock.
-          ) as unknown as AbstractTokenPricesService['validateCurrencySupported'],
+          validateCurrencySupported(_currency: unknown): _currency is string {
+            return false;
+          },
         });
         nock('https://min-api.cryptocompare.com')
           .get('/data/price')
@@ -2288,11 +2286,9 @@ describe('TokenRatesController', () => {
               value: 0.002,
             },
           }),
-          validateChainIdSupported: jest.fn().mockReturnValue(
-            false,
-            // Cast used because this method has an assertion in the return
-            // value that I don't know how to type properly with Jest's mock.
-          ) as unknown as AbstractTokenPricesService['validateChainIdSupported'],
+          validateChainIdSupported(_chainId: unknown): _chainId is Hex {
+            return false;
+          },
         });
         await withController(
           {